requirements.txt
catboost==1.2.1
numpy==1.25.2
pandas==2.1.0
lightautoml==0.3.8b1
transformers==4.32.1
torch==2.0.1
torchvision==0.15.2
torchaudio==2.0.2
scikit-learn==1.3.0
Pillow==10.0.0
pytorch-tabnet==4.1.0
matplotlib==3.7.2
seaborn==0.12.2